completed separation of component
