+++ public/index.html
@@ -3,7 +3,8 @@
   <head>
   </head>
   <body>
-    <div id="businessInfo">replace me</div>
+    <div id="businessInfoHeader">Biz info header goes here</div>
+    <div id="businessInfoSidebar">Biz info sidebar goes here</div>
     <script src="./bundle.js"></script>
   </body>
 </html>
